Logging

+++ DX11Renderer/DX11Renderer/Graphics.cpp
@@ -110,8 +110,8 @@ Graphics::Graphics(HWND hWnd, int windowWidth, int windowHeight)
 	//
 	// (includes screen size, but could be sub-portion of screen too)
 	D3D11_VIEWPORT vp;
-	vp.Width = windowWidth;
-	vp.Height = windowHeight;
+	vp.Width = (FLOAT)windowWidth;
+	vp.Height = (FLOAT)windowHeight;
 	vp.MinDepth = 0;
 	vp.MaxDepth = 1;
 	vp.TopLeftX = 0;
